misc/HtmlTestSuite/results/test-0644.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS UI Level 3: parsing outline with invalid values</title>
<link href="mailto:ericwilligers@chromium.org" rel="author" title="Eric Willigers"/>
<link href="https://drafts.csswg.org/css-ui-3/#outline" rel="help"/>
<meta content="outline supports only the grammar '&lt;outline-color&gt; || &lt;outline&gt; || &lt;outline&gt;'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_invalid_value("outline", "invert solid rgba(10, 20, 30, 0.4)");
test_invalid_value("outline", "double invert groove");
test_invalid_value("outline", "thin outset thick");
</script>
</body>
</html>
